\name{Clonality-package}
\alias{PairSeek-package}
\alias{PairSeek}
\docType{package}
\title{
Clonality testing
}
\description{
Flexible statistical method for identifying prognostic pairwise relationships among bacterial species in microbiome studies

}
\details{
\tabular{ll}{
Package: \tab PairSeek\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2020-12-20\cr
License: \tab  GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{

Irina Ostrovnaya \email{ostrovni@mskcc.org}, Sean Devlin \email{devlins@mskcc.org}
}
